refactor sequelize model loader and service
