+++ src/config/sequelize.config.ts
@@ -1,5 +1,5 @@
 import { Sequelize } from "sequelize";
-import modelLoader from "../utils/model-loader";
+import modelLoader from "../utils/sql-model-loader.utils";
 
 // Constants
 const HOST = process.env.MYSQL_DB_HOST || "localhost";
@@ -18,18 +18,3 @@ export const sequelize = new Sequelize(SCHEMA, USER, PASSWORD, {
 });
 
 modelLoader(sequelize);
-
-// Authenticate to the Database
-export const authenticate = async () => {
-  try {
-    await sequelize.authenticate();
-    console.log(
-      `Connected to the database at: ${HOST}/${SCHEMA} with user: ${USER}.`
-    );
-  } catch (error) {
-    console.error(
-      `Unable to connect to the database at: ${HOST}/${SCHEMA} with user: ${USER}.`,
-      error
-    );
-  }
-};
