+++ src/middleware/authenticate.middleware.ts
@@ -1,6 +1,6 @@
 import { NextFunction, Request, Response } from "express";
 import jwt from "jsonwebtoken";
-import { sendErrorResponse } from "../utils/responses.util";
+import { sendErrorResponse } from "../utils/responses.utils";
 
 // Constants
 const SECRET_KEY = process.env.AUTH_SECRET_KEY || "";
