+++ src/routes/auth.routes.ts
@@ -3,7 +3,7 @@ import bcrypt from "bcrypt";
 import jwt from "jsonwebtoken";
 
 import { User } from "../models/mysql/user";
-import { sendErrorResponse } from "../utils/responses.util";
+import { sendErrorResponse } from "../utils/responses.utils";
 import { Role } from "../models/mysql/role";
 
 // Constants
