+++ src/routes/contact-info.routes.ts
@@ -1,6 +1,6 @@
 import express from "express";
 import { ContactInfo } from "../models/mysql/contact-info";
-import { sendErrorResponse } from "../utils/responses.util";
+import { sendErrorResponse } from "../utils/responses.utils";
 
 const router = express.Router();
