+++ src/routes/paper-types.routes.ts
@@ -1,6 +1,6 @@
 import express from "express";
 import { PaperType } from "../models/mysql/paper-type";
-import { sendErrorResponse } from "../utils/responses.util";
+import { sendErrorResponse } from "../utils/responses.utils";
 
 const router = express.Router();
