+++ src/routes/subscriptions.routes.ts
@@ -1,6 +1,6 @@
 import express from "express";
 import { Subscription } from "../models/mysql/subscription";
-import { sendErrorResponse } from "../utils/responses.util";
+import { sendErrorResponse } from "../utils/responses.utils";
 
 const router = express.Router();
